LPS-124200 Rename and remove not necessary attributes
